improve setup.py and extern build configuration
Cleaned up setup.py and started it from the ground up.
Added a setup_helpers module to contain the under-the-hood
code needed by setup.py.
We extend the build and clean Commands from setuptools and
numpy's distutils.

This commit is not stable yet, but is being pushed upstream
so that I can test it on different platforms.