fix and improve the autotools configuration
In particular:
WCSSUBS is not included in the source distribution (make dist) of the whole
sherpa-dependencies package.

Autoconfiscated WCSSUBS by wrapping it in the main Makefile.am. This way
WCSSUBS's build is not ad-hoc anymore, and it supports all of the automake
features, avoiding pesky issues I was experiencing during development.

GROUP python extension is not created from the GROUP's python makefile anymore,
because the Makefile wasn't portable, and we need python to be called with the
same configuration as sherpa. The code is built as an extension directly from
sherpa's setup.py

Updated main configure.ac to get rid of some useless code generated initially by
autoscan and to support sherpa's configuration options for whether or not some
dependencies should be built or taken from the system.

Updated region's Makefile.am to get rid of obsolete clauses for orphan code.