fix and improve the setup.py script and its helpers
In particular:
setup now catches errors in the external code build, so that user is immediately
notified of errors (python wouldn't complain but the final result would be
junk).

added python dependencies into install_requires. This includes fixing an issue
with the way setuptools and numpy monkeypatch each other and distutils, in
particular for the sdist command: setuptools would produce a distribution that
would fail to build the f2py wrappers making the build of the fortran extensions
fail. For this reason I picked the numpy's sdist.py file and edited to make it
monkeypatch distutils's sdist instead of setuptools'.

notice that having setuptools was required if we want to use install_requires,
that instructs tools like easy_install and pip to download and install the
python dependencies.

added a setup.cfg file for managing the options of the newly created
sherpa_config command. This is standard distutils stuff, not the ad-hoc
solutions sherpa was employing before. As such we take all of the
standardization benefits for free.

setup.py is now very clean and simple, and sets the high level static metadata.
The dynamic configuration of extensions happens in setup_helpers, which provides
a setup wrapper function, inspects the options, and dynamically creates the
options to be passed to setuptools.

Sherpa also creates a file in the extern directory so that it only builds the
external dependencies successfully once, and then requires a
python setup.py clean (--all) to get rid of it. The alternative would have been
for the external code to be configured too often, even if it wasn't rebuilt.

a new MANIFEST file makes sure that all the files that should end up in the
source distribution are included.

setup.py was rewritten from scratch and it does not try to mess with the
compiler flags or to hardcode the compiler paths, so to maximize
interoperability and platform agnosticism, while leaving users control over
their setups. For instance, a simple CC='mycompiler' CXX='myc++' are enough to
point to a different compiler, and the other distutils/setuptools hooks can be
used to do more advanced stuff. Same goes for the compiler flags, which might
depend on the compiler, making the complexity grow and with it the possibily of
failures.