tweak several build-related files for milestone 1.7 (sherpa as OTS in CIAO)
In order to make sure that sherpa can be built both as a standalone package and
as OTS for CIAO I needed to tweak several autotools files, and several
helpers functions/classes. New options were added or tweaked in setup.cfg,
which is the only interface element changed in this commit.