add workaround to rebuild issue with fftw
Unfortunately, despite maintainer mode is disabled by default
in fftw, for some reason the rebuild rules are triggered when pip
drops the timestamps when unpacking tars.

The workaround is to "manually" touch fftw3.info so that
'missing' does not start to look for makeinfo, which is not
required for users.