Change Y label of the residual plot.
This changes the label of the Y axis for residual plots (e.g.
plot_resid() and plot_fit_resid()) to say 'Data - Model' rather
than 'y' for simple (e.g. Data1D) data types.

The existing behavior for DataPHA datasets - which is to display
the data units - is not changed. The code that enforces this is
not ideal, and I think that the label choice for DataPHA objects
should be reviewed. I also think that this highlights a
potential short-coming in the current plot design (i.e. how to
overload the basic plot settings, but this is not addressed
in this patch set.