Use LaTeX display in matplotlib for plots.
The plot_chisqr() command now uses the LaTeX support in matplotlib
for the title and Y axis labels (previously the text '\chi^2' was
displayed).

There are other places where LaTeX support could be added - for
instance, in DataPHA objects, the dependent axis is often displayed
with units of photon/cm^2/s/keV, so the '^2' could be made an
actual subscript. However, with the current design I don't see how
this can easily be done.