Minor test clean up.
It's not obvious why a failed test was being reported via the logger
and not the assertion. Possibly due to recent test changes the logger
isn't active when running this test, so no useful output (i.e. the
name of the model that failed) was being recorded if a test failed.