Initial fix for #62.
The crash occurs because the "gap" bins are calculated by a call
to the model with only 1 bin (so 2 elements for the input energy
array). This fix is to change this to 2 bins (so 3 elements), and
ignore the result for the second bin.

This does allow many methods to now be called with a non-contiguous
bin. However, the XSnsagrav model causes a crash when used in the
test case. I am reluctant to blame this on the nsagrav XSpec code,
as I am not convinced that there aren't other issues with this
interface code. I also plan to try a version whereby extra bins
are added into the input grid for non-contiguous arrays, so that the
function is only called once, to see whether this improves things.