Ensure XSPec models return 0's in case of error.
The XSpec model API does not provide a way of indicating that a
model failed. The common case is that an error message is printed,
often at a chatter level greater than the 0 used by default by Sherpa
(so that users do not see it) and the routine returns without filling
in the flux array. The flux (and error) arrays are now created so that
there contents are set to 0 rather than being undefined. This provides
a more-consistent, and more-obvious, signal to the user that a model
has failed (the result is 0).