Change spectrumNumber/ifl from 0 to 1.
The XSpec "spectrum number" value is not used by Sherpa. It was set to
0, but this has been seen to cause strange behavior in an XSpec user
model, so the setting has been changed to 1 to avoid this issue (it
is not clear, but likely that XSpec numbering for this field starts
at 1).

I also added some comments on how this value could be set by the user,
in the hope of supporting the smaug user model, but this is a long-term
change and may not work out as currently documented.