Fix XSpec FORTRAN float* models with non-contiguous wavelength.
For XSpec FORTRAN 4-byte real models only, change how the conversion
from Python input array(s) - in Sherpa "format" - to that needed by
XSpec (the n+1 length ear array and n length flux/fluxerror arrays)
is done. The old approach was to identify gaps in the array, call
the XSpec model with the input grid, and then loop over each gap,
calling the XSpec model with the correct bin width, over-writing the
value in the output array. The new approach is to call the XSpec
model once, with extra energy bins added for each gap, and then
removing these bins when forming the output array. The idea is that
this minimises the number of calls to XSpec, which may be an
optimisation; it may also "help" with those models that interpolate
answers onto the output grid, as well as being a more-robust solution
to issue #62. I have not really tried to optimise the code, preferring
readability to complex loops; it may be worth looking at the run time
cost once an approach has been settled upon.

A result of the above changes is that the bug in the handling of gaps
in wavelength grids has been fixed. This bug is described in
https://github.com/sherpa/sherpa/issues/62#issuecomment-120963426
but note that the model used in this example is a C++ style XSpec
model, so it will still show this behavior. Use of a model like
XSbbody or XSapec will show a change in behavior (but without the
advantage of having a Sherpa-supplied model to compare against).

The tests have been changed to reflect the code changes: more models
are tested with non-contiguous bins given in wavelength units (the
Fotrtan 4-byte models), but some tolerances have had to be relaxed
(for the wavelength checks) since it appears that the change in the
code has made some nunerical behavior worse (e.g. the conversion from
energy to wavelength is now down with 4-byte reals rather than 8-byte
reals which is then downcasted to 4-byte).