Use vector not malloc in XSpec template models [skip ci]
For some reason the template for XSpec table models was using
direct memory allocation (malloc/free) rather than just use a
vector for the energy array sent to XSpec models. This is unlike
the other templates. This commit switches to using a vector as
there is no obvious reason why this is needed.