Update convolution model interfact
This updates the interface for the convolution models - provided by
the xspecmodelfct_con template - to match the approach used by the
C++ template introduced in 6f1884ea58e3f59fd5c83dc70a4a59f29e1d1076

The major difference is that non-contiguous grids are not supported
by this model. The code has been updated to reflect this (but perhaps
it should have just retained the other code, but fixed ngaps to 0
so that the optimiser could have avoided unnescessary work, just to
keep the code the same - or as much so as possible).

A test has been added to check that convolution models fail if sent
a non-contiguous grid.