Fix error handling and code cleanup
The main changes are to correctly catch errors, although it is not
clear if the OutOfBoundErr class is ever raised (perhaps from compiled
code?). There are several other changes: removal of unused variables,
chamge min/max variable names (not all functions are changed),
comments on screen output that maybe should use the logging output,
more Pythonic boolean checks (e.g. change 'False == xxx' to 'not xxx'),
and a couple of PEP8 adding-space-around-symbol changes that accidentally
made it in.