Fix missing test
The test_ui.py tests can now be run directly with no test directory. This
revealed that there was a missing test in the test_psf_ui class, due to
what appears to be a typo.

Note that running this test directly with a directory - e.g.

    python sherpa/ui/tests/test_ui.py sherpa-test-data/sherpatest

will fail because the model name - which appears in an error message
that is being checked - is different (expect pl, get xspowerlaw.pl),
which is presumably down to the different environment that the test
runner uses compared to running the tests separately. This behavior
is not down to the change in this commit.