Allow test suite to run with a partial Xspec model build.
@olaurino asked if the Xspec test suite could be run with a partial
Xspec build (i.e. a subset of models), so this change should help
this (it also supports running with different versions of the Xspec
model library, which may add or remove models).

There's also a few minor white-space changes and commenting-out of
unused variables.

This commit is mainly to test out the new XSpec Travis-CI build, but
is still a meaningful commit.