Ensure XSPEC model evaluation is finite.
Following a discussion with Keith Arnaud, who says that XSPEC checks that
models evaluate to a finite value (to guard aainst errant user models from
crashing code), I have added such a check to the model code. This could
have been place in the template code (in _xspec.cc), but I placed it in
the XSModel class instead because:

a) it's simpler to write

b) it allows a user to directly call the model (via the _calc attribute
   or by directly importing _xspec) and find out what bins are evaluating
   to a non-finite value. This might be useful in debugging.

As the model evaluation now checks for finite values, the test code
no-longer needs to do so (the removed code was re-written slightly in
PR #63 to something a bit nicer, but this change superceeds that).