Allow writing of datasets with commentary cards.
The main change here is that the conversion from Sherpa header information
to astropy.io.fits format has been centralised and support added for
so-called commentary cards - COMMENT, HISTORY, and blank keywords - whose
values are an array of strings.

Tests need to be added, since at present the set_xxx_data routines are
not tested.