Add test for writing out PHA dataset.
This is a basic test of writing out a PHA dataset (it writes it out,
reads it back in, then compares several important values). It is not
complete (in the sense that there are more values that could/should
be tested), but I want to see how this works with the crates backend
before extending it (only tested against astropy).