Only load sherpa.astro.io if FITS support is on
Re-work the test to only load sherpa.astro.io if FITS support is
available. An easy way would be to only run the tests if there
is FITS support, doing nothing if there isn't, but I wanted to
keep the knowledge that tests had been skipped, hence the approach
used here. There may be a better way to do this.