Bugfix (undefined symbol) and PEP8 change.
In preparation for moving the tests to sherpa/astro/ui/utils, I did
some simple PEP8 fixes, and found an undefined symbol error. This is
unlikely to have caused serious problems, since it was in an error
case (if the tuple function raised a TypeError). The symbol
_argument_type_error is now imported from sherpa.ui.utils.

The PEP 8 fixes are:

- explicit imports
- replace `a == None` by `a is None`
- replace `a == True` and `a == False` by `a` and `not a` respectively
- replace `v.has_key(k)` by `k in v`
- remove some un-used variable assignments

Excess spaces on lines, and un-necessary brackets around some bool
conditions, have been removed.