Move tests sherpa.astro.io to sherpa.astro.ui
Since there are issues with importing sherpa.astro.io when there is
no FITS support available (i.e. no crates or astropy/pyfits), move
the tests to sherpa.astro.ui. This means that tests are now added for
the higher-level routines, but fortunately they are "almost" a drop-in
replacement for the lower-level routines (a little bit of plumbing is
needed). This is not ideal, but is a short-term fix until we decide
how to handle importing sherpa.astro.io when no backends are present.