Bug fix: save the background grouping and quality arrays.
Refactored code for serializing PHA files and fixed what appears to be
a bug, where the grouping and quality arrays for background files were
being taken from the source data set instead. This is only an issue for
people who fit the background, rather than subtract it (since in the
latter case the backgroun grouping and quality arrays are ignored).

The Python serialization is also a bit-more explicit about the optional
function arguments, in that the form name=value is now used in several
places related to PHA data files, rather than relying on position.