Save simple user models.
The serialization of user models requires that the inspect module can
access the code definition of the function called by the user model.
Even if it can not, the load_user_model and add_user_pars are included
to make it easier for any user to add them back in.