Add tests of fitting a model to a PHA background dataset
The output of save_all is not correct in this case, which means that
the output can not be run (the model expression for the background
component is not written out correctly). The test checks for the current,
incorrect, output. The "restore" test - which loads in the output of
save_all - is skipped for now.