bug: save background model expressions
The background model wasn't being queried correctly (incorrect use
of arguments) and so could be written out as using set_bkg_full_ model
(which does not serialize correctly for PHA data sets at this time)
instead of set_bkg_source. This was seen for data sets with a non-default
id (or if multiple backgronud components).