Add test of XSPEC table model
Adds a basic test of the XSPEC table model support: it does not include
wavelength or non-contiguous grids as these currently fail.

Added a test that the norm parameter of XSPEC additive models works,
since this is handled differently to the other XSPEC model parameters
(in that Sherpa code multiplies the return value from the XSPEC model
library by the normalization, whereas the other parameters are sent
to the XSPEC model library).

When evaluating the models, compare the results from a single argument
to that obtained with two arguments (i.e. low/high bin edges). This
is for both energy and wavelength grids, but is only for the contiguous
case (i.e. there are no gaps in the grids).

Add a test that the model will fail if the input grid is too small.

Minor clean up of the existing test code.