Separate out the convolution-model interface
The interface to XSPEC convolution models is currently experimental,
and no Python model classes are currently available (so you can
not make an instance of the model using the sherpa.astro.ui
module). The interface is not documented because of this.

This commit makes a separate template for the convolution models,
rather than shoe-horn it into the C/C++ model template. This change
simplifies the latter template. The interface for the convolution
models has changed from pars, xlo, fluxes, and optional xhi to
pars, fluxes, xlo, and optional xhi arguments. This is a breaking
change, but as discussed above the interface has not been documented.

Added simple tests of the convolution interface.

Added support for the cpflux convolution model.