Remove XSPEC sirf model from the tests
The XSPEC sirf model - up to 12.9.0d at least - is known to cause
occasional test failures, so remove it from the test suite. The
problem has been reported to the XSPEC developers - there is a
read from one element past the end of an array - which will hopefully
be fixed during the XSPEC 12.9.0 patch series.