Fix handling of non-contiguous and wavelength grids
This commit fixes problems seen

- XSPEC 12.8.2 using models that call the XSPEC SUMDEM routine (e.g.
  many thermal models, such as apec) when the grid was not contiguous,
  where a crash could result

- fix handling of non-contiguous grids when wavelengths are being used

When handling non-contigous grids with ngap gaps, the previous approach
was to call the model 1 + ngap times: the first time for the full grid
ignoring the fact that some bins were too large, and then once per gap,
using these values to fill in the incorrect results. The new approach
is to just call the model once, but with extra bins added to the energy
array, with the extra bins being removed before returning the results to
the user.

The code has also been re-written significantly to be clearer, and to
try and use the same code in all templates (re-factoring this is
out of scope of this PR). The code may not be as fast, since it is
written for clarity rather than speed, but

- no performance testing has been attempted
- hopefully the compiler and optimiser can recover possibly-faster versions

The models that work with 32-bit floats retain the old behavior of
working with a 64-bit energy grid, converting to 32-bit just before
calling the XSPEC model function. This conversion *could* be done
by `PyArg_ParseTuple` - i.e. when the data is converted from Python to
C/C++ scope - but this was not done since previous experiments have
suggested this can cause user confusion (since the grid is used to
identify non-contiguous bins, a source expression containing both
32- and 64-bit float models could end up using a different grid).
A number of checks - to ensure that the grid is monotonic and there
are no overlapping bins - have been commented out since there are
cases where the energy grid defined by the instrument response (i.e.
ARF or RMF) doesn't match these constraints (sometimes due to
32/64-bit conversions). They can be added back in at a later time, if
found to be useful.

The test suite has been enhanced to test models with both energy and
wavelength grids - using both single and double argument forms - and
to test bot contiguous and non-contiguous behavior. The tests do catch
the old, broken, behavior (i.e. they fail if used without the
code changes in this commit).