Fix NameError bug
The _argument_type_error routine, used to report a type error, was
not defined. This would have resulted in a confusing error message
but would only be seen when there was an error.

A number of PEP 8 fixes have also been made, including

- change 'x.has_key(y)' to 'y in x'
- change 'x == False' to 'not x'
- change 'x == None' to 'x is None'
- add spaces around operators ('x+1' -> 'x + 1') and after commas
  ('1,2,3' -> '1, 2, 3')
- explicitly import all symbols
- remove some un-used symbols
- remove excess brackets in conditional statements
- avoid line overlaps

The save_all/save_session routines have not been touched as there are
existing PRs that change this code.