Fix writing out of FITS files with commentary cards
The code now handles FITS "commentary" cards - e.g. HISTORY, CONTINUE,
and blank - which can have multiple values. The code was refactored to
use a common routine for converting between the Sherpa metadata format
and that used by astropy.

The decorator for the testWrite test marking it as expected to fail
has been removed.