Support .gz PHA files in astropy
Support reading PHA files with and without .gz extensions using astropy.
This hopefully matches the crates behavior - e.g. if pha.fits.gz exists
but pha.fits does not then load_pha('pha.fits') will load the .gz form.

This is particularly useful when loading Type 3 PHA files - as produced
by the Chandra Source Catalog.

This change includes an update to the sherpa-test-data submodule. It
includes several unrelated changes that are in other patches that have
yet to be merged.