Documentation updates
There are no code changes here, only documentation updates (and moving the
copyright statement to the start of the file). The documentation changes
include a major expansion of the module-level docstring, addition of
various functions, and updating to numpy-style formatting. The new text
is based on the CIAO ahelp documentation for the datastack module.