No longer replace docstrings of Sherpa functions.
The following applies to the CIAO-supplied ipython customization
routines used by the "sherpa application" in CIAO (which is a thin
wrapper around ipython). It does not change the behavior of the
stand-alone build of Sherpa.

Prior to this change, routines from sherpa.astro.ui had their
docstrings replaced with the text "Type ahelp(func) for more deails.",
where func was replaced by the function name. As the docstrings have
now been updated to contain the ahelp text, this replacement is
not needed. Note that the ahelp() function is not removed, so
users who enter `ahelp(func)` will get the ahelp message displayed.