Remove the check for non-finite values from XSPEC models
The XSPEC interface changes in Sherpa 4.8.0 development included a
change where the sherpa.astro.xspec.XSModel class would check that the
model had created only finite values, raising an error if this
constraint did not hold. This was suggested as a good thing to do
during a meeting between the Sherpa developers and Keith Arnaud.

This change caused problems with some of the external regression tests
in CIAO, which use a problematic (i.e.  does not confirm to OGIP
standards) RMF file with an energy bin that includes 0 keV. This
causes the XSpowerlaw model to create a non-finite value in this bin
when the photon index is positive.

This commit reverts this change (by commenting out the code that added
the check), and adds in explicit checks in the test suite.  Hopefully
the check for non-finite values can be added back in once the handling
of this case has been reviewed.