Run load_table_model tests from sherpa.ui against sherpa.astro.ui
All-but two of the load_table_model tests from sherpa.ui.tests.test_ui
are run against the sherpa.astro.ui version in
sherpa.astro.ui.tests.test_astro_ui. This is done by sub-classing the
test suite, which requires adding a __init__.py file to the sherpa.ui.tests
directory. Is there a better way to do this?

For some reason the test_fail_on_missing_col test does not raise an IOErr
with the sherpa.astro.ui version. This may be because of support for
XSPEC table models changing the interface, or something else.

This commit does not extend the tests to include FITS files for the
sherpa.astro.ui version.