Separate out load_table_model
Does it make sense to have these "helper" routines members of the Session
class, rather than just utility functions?

An Error has been changed from Exception to IOErr, to better match the
other parts of Sherpa. Note that this routine is only called in one place,
which checks for any exceptions thrown, so this is not a functional change,
but now that _load_xspec_table_model has been separated out it makes more
sense for it to act more like other parts of the system.