More precise testing of a test
This checks that when a table model can not read in a column the error
message is as expected. This allows the failure case with the FITS support
to be labelled `expectedFailure` now (previously the test passed because
it was not picking up the "correct" failure).

Minor comment and documenation updates.