Fix bug #177
This updates several tests to check for numeric equivalence using a
tolerance rather than absolute equality. There have also been PEP8
changes, and updates to use more-recent unittest methods (e.g.
assertTrue rather than assert_ and assertGreater).

Several tests should probably be re-written to use vectorized routines
rather than iterating over elements, but that is outside the scope of
this fix.

The problem test could be re-written to use a different numerical
operation, which could remove the need for a tolerance check, but I
felt that was also out-of-scope here.