More explicit file I/O
The sherpa.io module has been enhanced by adding a routine that will read
in column data, but not force it to match the number of arguments needed
to create a Data object. This routine is then used in sherpa.astro.ui.utils
to read in data in the fall-back path that the sherpa.astro.io routines
have failed.

This change required having some way to know - in sherpa.astro.ui.utils -
whether the sherpa.astro.io module was available (i.e. whether the crates
or astropy backend is loaded). It is not clear to me what the best way
to do this is.

The change also loses potentially-useful error messages - e.g. instead of
something like "Unable to find column 'x' in ['a' 'b']" it is now just
"No input array(s) found".

These changes fix the test_astro_ui.py tests when no astro I/O backend
is available.