Protect XPA command to avoid shell confusion.
This fix is to avoid `sherpa.astro.ui.image_data()` from failing when
the working directory contains a file called `x` or `y` (and possibly
other combinations). It is a limited fix; ideally the code should be
updated to take better advantage of the `subprocess` module (e.g.
using a sequence rather than a string for the arguments, so that
there is no need to "protect" arguments).