generalize Stat API
When I was asked to pass the bkg data to the UserStat, I made a mistake by
making the code backwards compatible.  Then when I had to add WStat to Sherpa,
things went from bad to very bad.  After all, WStat needs the exposure time and
backscale ratio.  Not to mention the size of the data sets if one is to use it
for fitting simultaneous data sets (different data sets could have different
exposure time and/or backscale ratios).

The old classes for the Stats are not very easily extensible; Moreover, The
changes should also enable the user to do simultaneous fit using different
statistics for different data set.