Reword the README
The testign section has been pulled out so that the same text can be
used for both binary and source installs (although it is not ideal, I
felt it was better separated in this way).

Added in a link to the TOC, and separated out the "hey, also install
these packages" as a new section.

Reworded some sections, including removing repeated information (due
to the changes above).

Updated the XSPEC build instructions to better match version 12.9.0,
as well as point out that the "use CIAO binaries" is not the
recommended approach.