Add documentation to model classes
The information added was based on a combination of the CIAO 4.8
documentation for the models (ahelp files), and an inspection of the
code. In particular, note that the Poisson model implements a
different function to that described in the CIAO 4.8 documentation
(there is an extra log(mean) term).

It is not yet clear to me how the best way to document the model
parameters is: I have used the "Attributes" section, since they are
attributes of the model, but there is currently no mention of their
type (e.g. the Parameter class), or other related attributes of the
model.

I use a version of ASCII maths, rather than embed LaTeX, to enable
them to be read at the command line. This could be changed.