Improve sherpa.astro.optical model documentation
Add in documentation for absorption models and some emission cases
I had not got around to.

Be more explicit about the units of the independent axis (for some
cases, such as the Powerlaw class, the units need just be in
wavelength as long as the reference parameter is taken to be in the
same units, but these parameters are labelled as taking Angstrom so
be explicit for them).

The LogEmission model has two issues:
  - unused limit parameter (#219); I have noted this is unused in the
    documentation
  - difference between docs and code (#220); I have documented the code
    rather than the equations given in the pecview documentation for now