Convert from basestring to six.string_types for string checking
In reality a lot of the isinstance checks used here aren't ideomatic
Python and should probably be changed, but that is outside the scope of
this work.

I have kept one instance of basestring, where it was explicitly imported
from numpy.compat (in sherpa/astro/io/pyfits_backend.py). This needs
reviewing.