Remove deprecated warnings from several uses of inspect.getargspec
I added tests to two of the functions, to catch the "all arguments are
keyword arguments" case, and a test for get_num_args.

It looks like inspect.getargspec is not being removed in Python 3.6,
but I think this change is still useful (in that it avoids deprecated
warnings from Python 3.5). I am not convinced it will work for Python
3 before 3.3, but I do not think we are planning support for those
versions.