More use of inspect.signature rather than inspect.getargspec
I added several tests to make sure some of the trickier aspects of the
conversion were tested.

The replacement of getargspec by signature in export_method does not
seem to be necessary, in that the old version did not cause deprecation
warnings to be displayed. It is not clear to me why that is, so this commit
may not be needed.

It also isn't clear to me what export_method is doing to need a version
of the function with the default values removed, but I don't want to
investigate this any further at this time.