Add a micro-optimisation and tests.
I prefer the approach here - that is, just looping through .pars
rather than checking everything in __dict__ - but it found a
problem case in the XSPEC models, which used the wrong case for
a model parameter during the initialisation call, which lead to
the messier code we now have. It did lead to extra tests.

The code for __getattr__ could be simplified to avoid the fallback
to looping over __dict__, but this would involve a policy change - i.e.
require that parameters are capitalised correctly when the model is
being initialised (i.e. during __init__ calls). That is, to outlaw
something like

    class MyModel(ArithmeticModel):
        def __init__(self, name='mymodel'):
            self.nH = Parameter(name, 'nH', ...)
            ArithmeticModel.__init__(self, name, (self.nh, ))