Add explicit tests for wstat
This test suite is intended to provide explicit checks for various
different ways that datasets can be set up for use with wstat. This
is to support development during the 4.8 release cycle, since ideally
these tests would be spread throughout the system (the tests here
are regression tests and do not provide substantive tests of the
logic of the wstat calculations).