Add a nightly build option
Python 3.6 is now close to release, so test out the nightly branch which is
hopefully close to this release: see
https://docs.travis-ci.com/user/languages/python/#Choosing-Python-versions-to-test-against
for more details.

I picked the Python 3.5 test as the one to copy, and marked it as
allowed-to-fail, as this is intended to be informational since we do
not claim Python 3.6 support yet.