Add new calc_stat_from_data method to stat class
This is the start of a rewrite to how the sherpa.fit.Fit class passes
data to the statistics object, with the aim of moving the "data selection"
logic - that is, what items the statistic object needs - from the fit
object to the statistics object.

A new method has been added to the sherpa.stats.Stat class, called
calc_stat_from_data. This routine is intended to replace the current
calc_stat (i.e. it will be renamed to calc_stat) but it is left separate
at present. Unlike calc_stat, it is sent Sherpa data objects (e.g.
classes following the sherpa.data.Data API) rather than arrays of data.

This commit fixes the wstat failures (e.g. issue #248), so that the
expectedFailure/xfail decorators/labels have been removed.