Add basic 2D tests and fora 1D integrated dataset.
Two of the wstat tests added here (working with background-subtracted data)
have been calculated using this PR as the code produces different results
to CIAO 4.8. However, it's not valid to run the code in this case, so
it really should error out. I have added these tests to flag up when the
behavior is changed.