Likelihood statistics now error out if input is background subtracted
This commit resolves the problem with calculating statistic values for
background-subtracted data using likelihood states, noted in changeset
6cc48221af71566192163361ed9181feccff531b and issue #289

The approach taken is to add in basic validation routines to the Stat
object and then override them as necessary (to try and reduce the
amount of boilerplate needed in the calc_stat_from_data method). Since
the calc_chisqr method calls calc_stat_from_data (for chi-square based
statistics), they automatically gain these checks.

One down side to moving the checks here is that, in a fit (where some
of the assertions are not going to change each iteration) these checks
are called each iteration and so incur an un-needed cost. However, it
is unclear how to identify which are going to remain unchanged and
which are not (e.g. for iterated schemes), and I have not tried to
measure the slow-down that these checks add (so it may not be a
significant problem).

There may be more places where checks in sherpa.fit can be moved
into the stats hierarchy. This commit also does not attempt to remove
the checks in fit.