Fix #294
Adds a simple test to check the format of the __str__ method of
the jdpileup model. This provides limited checks that the information
is correct, but does test whether #294 is fixed.

The test was run before and after fixing `sherpa/astro/models/__init__.py`
to check that it failed and then passed using Python 3.5.