Added load_xstable_model routine.
To idenitfy cases where this routine is called but XSPEC support is
not available, a new type of sherpa.utils.err.ImportErr has been
created.

A number of tests are repeated, even though they could be parameterized
by the load function, since in this release both load_xstable_model and
load_table_model can be used. There are two reasons for the repeated
code:

- there are subtle differences (e.g. in what Exception is raised) which
  makes sharing code a bit messy
- I was unable to get the tests to correctly skip tests when necessary
  (e.g. if XSPEC support was not available).