Build a wrapper for XSPEC nlapec model if it exists
This commit just makes the building of a Python wrapper function
around the nlapec model (added in XSPEC 12.9.0, so not in XSPEC 12.8.2)
optional. It does not include the Python class for this model.

There does not appear to be a nice, clean, way to do build-time
detection of the presence (or absence) of a model in the XSPEC library.
This is in part because there's no guarantee that the runtime environment
is the same as the build time one (i.e. you can build Sherpa against one
version of XSPEC and then point to a different one when you use Sherpa).
It it also because the two ways to check if the model are present are:

1) check it is present in libXSFunctions, but it is not just as simple
   as using ctypes.cdll.LoadLibrary(<path to this>) since it requires
   symbols from other modules

2) parse the model.dat file and look for it there, but the problem with
   this approach is that (at present) there is no requirement for this
   file to be present when building the XSPEC module in Sherpa.

This patch goes with option 2, by guessing where the model.dat file
is (if it can't be found then the nlapec model is assumed to not
exist).