Fix over-zealous code clean up in PR #287 affecting sigmarej
Fits using the sigmarej iterated-fit method were broken if a filter
had been applied to the data before the fit and there are any bins
that get ignored at larger bin values than the filtered-out data.

I added a test, checked that it failed before fixing the code and
fixed after the code. This deals with the "while not(newmask...)" line
change.

I manually checked the change for the "if current_par.frozen:" change,
by inserting errors in both branches and checking the test suite errored
out.

The last change - "not(any(newmask))" - does not have any test as it
is for an error condition and it is not obvious how to create a test
that triggers this.