Strip out AstroPy clobber deprecation warning
Remove AstroPy deprecation warnings from version 1.3.0 for the two
tests, since they are not intended to check for this warning. There
are also small changes to the test (moving some code out of the
warning path since it's really only save_data that is being tested
here), as well as pep8 tweaks.