Initial AREASCAL improvements
These changes improve handling of PHA objects with AREASCAL values that
are not unity. There are still unsolved problems

 - if backgronud is subtracted then the answers are still wrong
 - it's not at all clear to me what happens with the likelihood stats,
   since applying the area scaling to the counts seems like the wrong thing
   to do in this case
 - it's not clear what to do with set_dep; should it "uncorrect" the
   supplied values by the existing areascal column (see comments added
   to this routine).

Something strange is happening with the two tests that check that a
RuntimeWarning is created when calling get_dep/get_y. The tests pass
when only the test script is run - e.g.

  python setup.py test -a sherpa/astro/tests/test_astro_pha_scale.py

but fail when all the tests are run:

  python setup.py test

The warning message checked for by these tests probably needs to be
changed to be version-specific (Python 2 versus Python 3).