Hide the divide-by-zero warnings
It's not clear that this is the best approach for the user, but in most
cases I would imagine they would not expect to see this warning if they
have ignored invalid points.

The reason for this change is to simplify testing, since the warnings
were interfering with the test suite, and I did not want to add this
divide-by-zero to the list of warnings to ignore because the current
scheme can not restrict the warnings to a single test.