First attempt at fixing errors when subtracting the background
This does not test that the area scaling is correctly handled when
other scales are present (primarily exposure time and BACKSCAL).

There are still differences when comparing background-subtracted error
values calculated with XSPEC, which appears to resolve around how bins
with 0 counts are handled (e.g. src cts=0 or bg cts). This appears to
be a generic issue, ie not related to AREASCAL handling, but more
investigation is needed.