Add basic tests of AREASCAL handling
The aim is to support the planned WSTAT rewrite (to support passing AREASCAL
to the WSTAT code), but other tests are added as well. At present this is
purely a regression-style test, since all that is being looked for is to
see if/when tests fail. This is because it's not obvious what the correct
approach to handling AREASCAL is, so the tests are being used to check
what is and isn't changing. The plan is to validate the test values once
things have settled down.