Support ARESCAL values for WSTAT statistic
This follows what I believe to be XSPEC's approach, following a private
communication with Keith Arnaud @xspector, which multiplies the exposure
time by the AREASCAL. This makes the source and background "exposure times"
become arrays, rather than scalars, which requires changing how the
information is passed into the C++ code. It actually makes things a bit-more
uniform, since the exposure times are now handled in the same way as source
and background counts and BACKSCAL values, rather than be special-cased.

The test is updated to reflect this change (that is
sherpa/tests/stats/tests/test_stats_unit.py::test_stats_calc_stat_pha_ascal)
but this is still a regression-style test, since it is known that the
data being sent to the WSTAT statistic is wrong when AREASCAL is set
(since it is being corrected in the DataPHA class when it should not be).