Add UI-level tests of CSTAT and WSTAT (comparisons to XSPEC)
Rather than add extra files to the sherpa-test-data repository, this
version adds in the necessary data to an existing file (on read). The
values computed are compared to those calculated with XSPEC 12.9.0o
(I don't expect the exact XSPEC version to matter too much since the
model is simple - a flat powerlaw - and hopefully the I/O and
statistic values aren't going to change too much).