Switch from using XSPEC power law to Sherpa for a test
Surprisingly enough this switch changed the stat value (from NaN
to a finite value). This value is only used as a regression test,
to ensure that we note when the behavior changes, rather than a
validation that the code is doing the right thing (which is later
on in this particular test).