Fix up agreement with XSPEC for background-subtracted data.
I had managed to over-confuse myself with the calculation of the
background-subtracted values. This commit should fix this.

There is some validation left, primarily a check that the plot values
agree between XSPEC and Sherpa (an external check, to validate the
tests of get_y and get_yerr in test_astro_pha_scale.py) and
validation of the "TODO: verify" numbers left in test_stats_unit.py.