Implement hide_warnings decorator; use it for DS9 tests
Adds a decorator - sherpa.utils.hide_warnings - that can be used to
hide particular warnings from a test. This is used to hide the
Python 3.6-specific warning messages from the DS9 tests in
sherpa/image/tests/test_image.py. For these particular tests the
warning list could have been made version specific (i.e. only
hidden for Python 3.6 or later), but it's not clear whether that is
a worthwhile restriction here.

This commit does not remove the "allowed to fail" flag on the
Python 3.6 test on Travis-CI.