Explicitly close the subprocess processes when calling ds9/xpa* tools
This is primarily to stop the ResourceWarning warnings that are created
when running sherpa/image/tests/test_image.py under Python 3.6.

In all-but-one case this is a simple fix, with the code being converted
to use the context-manager approach to subprocess.Popen. These are
the calls to xpaset and xpaget. The remaining issue is running ds9,
which is meant to be run in the background: without explicitly using
os.fork - which is not available on Windows - it's not clear how best
to handle this so I have used a hack. It might be sensible to store
the Popen instance when calling ds9 and storing it in the DS9Win class,
but it's not clear if this helps this particular issue.

This change means that the hide_warnings decorator is no-longer being
used, but I have left it in the code base for now.