Use a line not span to draw horizontal line (fix #378)
My guess is that the handling of zero-width spans has changed between
version 1.5 and 2.0 of matplotlib. It is also not clear why axhspan was
being used in the first place, since axhlie is used in another part of
the module.

There is an attempt to make the line match the axis (by checking what the
linewidth is), since without this change it is visually distracting in
matplotlib 2.0.