Add test to show the failure with python 3
This test is to show the existing failure. It requires a copy of the
session test for the astro version. It would be better if the common
code in these tests were factored out rather than copied!

The copy also means that more tests are used than are needed for this
bug, but if we are going to replicate tests on the two session objects
it makes sense to repeat all of them (that make sense).